fix: error number in dashboard

+++ database/migrations/1697032897916_clients.ts
@@ -9,7 +9,7 @@ export default class extends BaseSchema {
       table.integer('user_id').unsigned().references('users.id').onDelete('CASCADE')
       table.string('name', 255).notNullable()
       table.string('month', 255).notNullable()
-      table.integer('somme').notNullable()
+      table.decimal('somme').notNullable()
       /**
        * Uses timestamptz for PostgreSQL and DATETIME2 for MSSQL
        */
